OG commit
This is a pretty barebones site, it uses bootstrap 4 for the layouts, but otherwise its pretty simple. I have a JS file that ganerates a random job title from a noun and verb and "types" it out in a cool animation. Other wise this is a simple portfolio site. Colour scheme and layout changes to come. I think i want the theme to have a dark grey/purple theme with fades but lets see how far this goes